[feature] support for ME scope
ME is a scope that allows read and write access to the data of the user
associated with the access token.

This includes:

  * Retrieving the complete User resource
  * Modifying all attributes of the User resource
  * Deleting the User resource
  * Revoking the access tokens of the User
  * Access the /me resource
  * Validate the access token

Note that this only works with a user-bound access token, i.e. a token with
scope ME that has been retrieved via client credentials grant CANNOT access any
user's data.